+++ lib/Relay/Relay.hpp
@@ -2,12 +2,13 @@
 #define RELAY_HPP_INCLUDED
 
 #include <Arduino.h>
+#include <stdint.h>
 
 class Relay
 {
 private:
     bool state;       //current state of the relay: 0: off; 1: on
-    unsigned int pin; //relay pin
+    uint8_t pin;      //relay pin
 
     bool ON;  //physical relay state: on.
     bool OFF; //physical relay state: off.
@@ -17,8 +18,8 @@ private:
     void setState(bool desired_state); //Sets the relay state.
 
 public:
-    Relay(unsigned int relay_pin, //Arduino pin that the Relay is connected to
-          bool mode               //Logic mode (default: high-active, i.e. high level means on).
+    Relay(uint8_t relay_pin, //Arduino pin that the Relay is connected to
+          bool mode          //Logic mode (default: high-active, i.e. high level means on).
     );
 
     void on(); //Turns the relay on.
